machines: Pull NIC settings dialog body into reusable component t
