+++ pkg/machines/components/nic.css
@@ -1,0 +1,3 @@
+.machines-network-list .pficon-pending {
+    margin-left: 0.5rem;
+}
