--- pkg/machines/components/nicEdit.css
@@ -1,8 +1,0 @@
-/* Standard dialog is too narrow and wide dialog is too wide */
-.nic-edit > .modal-dialog {
-    max-width: 60rem;
-}
-
-.machines-network-list .pficon-pending {
-    margin-left: 0.5rem;
-}
